admin management

+++ backend/models/base.py
@@ -407,7 +407,7 @@ class TransactionStatus(str, Enum):
 class Wallet(BaseModel):
     id: str = Field(default_factory=lambda: str(uuid.uuid4()))
     user_id: str
-    balance_coins: int = 0  # Balance in coins (1 coin = ₦100)
+    balance_coins: float = 0  # Balance in coins (1 coin = ₦100)
     referral_points: int = 0  # Points earned from referrals and activities
     created_at: datetime = Field(default_factory=datetime.utcnow)
     updated_at: datetime = Field(default_factory=datetime.utcnow)
@@ -417,8 +417,8 @@ class WalletTransaction(BaseModel):
     wallet_id: str
     user_id: str
     transaction_type: TransactionType
-    amount_coins: int  # Amount in coins
-    amount_naira: int  # Amount in naira (for display)
+    amount_coins: float  # Amount in coins
+    amount_naira: float  # Amount in naira (for display)
     status: TransactionStatus = TransactionStatus.PENDING
     description: str
     reference: Optional[str] = None  # Payment reference/proof
@@ -434,8 +434,8 @@ class WalletFundingRequest(BaseModel):
     proof_image: str  # Base64 or filename of payment proof
 
 class WalletResponse(BaseModel):
-    balance_coins: int
-    balance_naira: int
+    balance_coins: float
+    balance_naira: float
     transactions: List[WalletTransaction] = []
 
 # Job with Access Fee
@@ -552,11 +552,11 @@ class VerificationSubmission(BaseModel):
 
 # Updated Wallet Response to include referral coins
 class WalletResponseWithReferrals(BaseModel):
-    balance_coins: int
-    balance_naira: int
+    balance_coins: float
+    balance_naira: float
     referral_coins: int           # Legacy field: Coins earned from referrals (calculated)
     referral_points: int = 0      # New field: Points balance
-    referral_coins_naira: int     # Referral coins in naira
+    referral_coins_naira: float   # Referral coins in naira
     can_withdraw_referrals: bool  # True if total coins >= 5
     transactions: List[WalletTransaction] = []
